refactor(ProjectService): Move consolidatePaths() to PathService
#890
