+++ src/assets/wise5/services/projectService.ts
@@ -943,7 +943,7 @@ export class ProjectService {
      */
     const pathsSoFar = [];
     const allPaths = this.getAllPaths(pathsSoFar, startNodeId);
-    const nodeIds = this.consolidatePaths(allPaths);
+    const nodeIds = this.pathService.consolidatePaths(allPaths);
     this.flattenedProjectAsNodeIds = nodeIds; // cache flatted node ids
     return nodeIds;
   }
@@ -1190,160 +1190,6 @@ export class ProjectService {
       pathSoFar.pop();
     }
     return allPaths;
-  }
-
-  /**
-   * Consolidate all the paths into a linear list of node ids
-   * @param paths an array of paths. each path is an array of node ids.
-   * @return an array of node ids that have been properly ordered
-   */
-  consolidatePaths(paths = []): string[] {
-    let consolidatedPath = [];
-    /*
-     * continue until all the paths are empty. as we consolidate
-     * node ids, we will remove them from the paths. once all the
-     * paths are empty we will be done consolidating the paths.
-     */
-    while (!this.pathService.arePathsEmpty(paths)) {
-      // start with the first path
-      const currentPath = this.getNonEmptyPathIndex(paths);
-
-      // get the first node id in the current path
-      const nodeId = this.pathService.getFirstNodeIdInPathAtIndex(paths, currentPath);
-      if (this.pathService.areFirstNodeIdsInPathsTheSame(paths)) {
-        // the first node ids in all the paths are the same
-
-        // remove the node id from all the paths
-        this.pathService.removeNodeIdFromPaths(nodeId, paths);
-
-        // add the node id to our consolidated path
-        consolidatedPath.push(nodeId);
-      } else {
-        // not all the top node ids are the same which means we have branched
-
-        // get all the paths that contain the node id
-        const pathsThatContainNodeId = this.getPathsThatContainNodeId(nodeId, paths);
-
-        if (pathsThatContainNodeId != null) {
-          if (pathsThatContainNodeId.length === 1) {
-            // only the current path we are on has the node id
-
-            // remove the node id from the path
-            this.removeNodeIdFromPath(nodeId, paths, currentPath);
-
-            // add the node id to our consolidated path
-            consolidatedPath.push(nodeId);
-          } else {
-            // there are multiple paths that have this node id
-
-            // consume all the node ids up to the given node id
-            const consumedPath = this.consumePathsUntilNodeId(paths, nodeId);
-
-            // remove the node id from the paths
-            this.pathService.removeNodeIdFromPaths(nodeId, paths);
-
-            // add the node id to the end of the consumed path
-            consumedPath.push(nodeId);
-
-            // add the consumed path to our consolidated path
-            consolidatedPath = consolidatedPath.concat(consumedPath);
-          }
-        }
-      }
-    }
-    return consolidatedPath;
-  }
-
-  /**
-   * Consume the node ids in the paths until we get to the given node id
-   * @param paths the paths to consume
-   * @param nodeId the node id to stop consuming at
-   * @return an array of node ids that we have consumed
-   */
-  consumePathsUntilNodeId(paths: any[], nodeId: string): any[] {
-    const consumedNodes = [];
-    for (const path of paths) {
-      if (path.includes(nodeId)) {
-        const subPath = path.splice(0, path.indexOf(nodeId));
-        for (const nodeIdInPath of subPath) {
-          if (!consumedNodes.includes(nodeIdInPath)) {
-            consumedNodes.push(nodeIdInPath);
-          }
-        }
-      }
-    }
-    return consumedNodes;
-  }
-
-  /**
-   * Remove the node id from the path
-   * @param nodeId the node id to remove
-   * @param paths an array of paths. each path is an array of node ids
-   * @param pathIndex the path to remove from
-   */
-  private removeNodeIdFromPath(nodeId: string, paths: any, pathIndex: any): void {
-    if (nodeId != null && paths != null && pathIndex != null) {
-      const path = paths[pathIndex];
-      if (path != null) {
-        for (let x = 0; x < path.length; x++) {
-          const tempNodeId = path[x];
-
-          /*
-           * check if the node id matches the one we are looking
-           * for
-           */
-          if (nodeId === tempNodeId) {
-            /*
-             * we have found the node id we are looking for so
-             * we will remove it from the path
-             */
-            path.splice(x, 1);
-
-            /*
-             * move the counter back since we just removed a
-             * node id. we will continue searching this path
-             * for the node id in case the path contains it
-             * multiple times.
-             */
-            x--;
-          }
-        }
-      }
-    }
-  }
-
-  /**
-   * Get the paths that contain the node id
-   * @param nodeId the node id we are looking for
-   * @param paths an array of paths. each path is an array of node ids
-   * @return an array of paths that contain the given node id
-   */
-  private getPathsThatContainNodeId(nodeId: string, paths = []) {
-    const pathsThatContainNodeId = [];
-    for (const path of paths) {
-      if (path.indexOf(nodeId) !== -1) {
-        pathsThatContainNodeId.push(path);
-      }
-    }
-    return pathsThatContainNodeId;
-  }
-
-  /**
-   * Get a non empty path index. It will loop through the paths and
-   * return the index of the first non empty path.
-   * @param paths an array of paths. each path is an array of node ids
-   * @return the index of the path that is not empty
-   */
-  private getNonEmptyPathIndex(paths: any): any {
-    if (paths != null) {
-      for (let p = 0; p < paths.length; p++) {
-        const path = paths[p];
-        if (path.length !== 0) {
-          return p;
-        }
-      }
-    }
-    return null;
   }
 
   /**
